Added percentages compared to the year before
